RoadDamageDetection-main/requirements.txt
av=10.0.0
numpy==1.24.3
Pillow==9.5.0
streamlit==1.28.0
streamlit-webrtc==0.47.1
torch==2.0.0
ultralytics==8.0.201
opencv-python-headless==4.8.0.74
requests==2.28.2